# typescript/openapi/jarvis.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Jarvis Operator API",
    "summary": "A safety-gated API for the Jarvis Z-State runtime and durable operator data.",
    "description": "Jarvis accepts structured operator requests, routes work through its runtime and safety envelope, persists tasks and reminders through the configured JSON or Convex provider, runs proposal-only Totality reasoning, stages revision-checked project-memory changes, and stores revision-bound tool-action proposals for explicit approval or rejection. Approved tool actions may be executed through the small reviewed server-side allowlist, with durable receipts and external reconciliation where required. The private development ChatGPT App preview provides a stateless Streamable HTTP MCP adapter and versioned dashboard widget for the operations marked with x-mcp-tool.exposed=true. Jarvis and OpenAI credentials remain server-side. This contract is intentionally single-user and private.",
    "version": "0.7.0",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "serviceToken": []
    }
  ],
...
  "paths": {
...
    "/api/v1/properties": {
      "get": {
        "operationId": "listProperties",
        "summary": "List business properties",
        "description": "Returns the authoritative property register, optionally filtered by client.",
        "tags": ["Properties"],
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "description": "Optional opaque client identifier used to filter properties.",
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ],
        "x-mcp-tool": {
          "exposed": false,
          "annotations": {
            "readOnlyHint": true,
            "destructiveHint": false,
            "openWorldHint": false,
            "idempotentHint": true
          }
        },
        "responses": {
          "200": {
            "description": "Current properties.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createProperty",
        "summary": "Create a business property",
        "description": "Creates one client-owned service property with address, access notes, service notes, and site hazards.",
        "tags": ["Properties"],
        "x-mcp-tool": {
          "exposed": false,
          "annotations": {
            "readOnlyHint": false,
            "destructiveHint": false,
            "openWorldHint": false,
            "idempotentHint": false
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Property created.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/properties/{propertyId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/PropertyId"
        }
      ],
      "get": {
        "operationId": "getProperty",
        "summary": "Get one business property",
        "description": "Returns one property by its opaque identifier.",
        "tags": ["Properties"],
        "x-mcp-tool": {
          "exposed": false,
          "annotations": {
            "readOnlyHint": true,
            "destructiveHint": false,
            "openWorldHint": false,
            "idempotentHint": true
          }
        },
        "responses": {
          "200": {
            "description": "Property found.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "patch": {
        "operationId": "updateProperty",
        "summary": "Update a business property",
        "description": "Updates the client link, address, hazards, access notes, or service notes of one property.",
        "tags": ["Properties"],
        "x-mcp-tool": {
          "exposed": false,
          "annotations": {
            "readOnlyHint": false,
            "destructiveHint": false,
            "openWorldHint": false,
            "idempotentHint": true
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Property updated.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "delete": {
        "operationId": "deleteProperty",
        "summary": "Delete a business property",
        "description": "Permanently removes one property by its identifier.",
        "tags": ["Properties"],
        "x-mcp-tool": {
          "exposed": false,
          "annotations": {
            "readOnlyHint": false,
            "destructiveHint": true,
            "openWorldHint": false,
            "idempotentHint": true
          }
        },
        "responses": {
          "200": {
            "description": "Property deleted.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Single-user Jarvis service token. Send it only in the Authorization header. Never place it in URLs, logs, tool arguments, widget state, or model-visible output."
      }
    },
    "headers": {
      "RequestId": {
        "description": "Opaque request correlation identifier safe to include in support reports.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        }
      }
    },
    "parameters": {
...
      "PropertyId": {
        "name": "propertyId",
        "in": "path",
        "required": true,
        "description": "Opaque property identifier returned by Jarvis.",
        "schema": {
          "$ref": "#/components/schemas/RecordId"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "The service token is missing, invalid, or revoked.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string",
              "example": "Bearer"
            }
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested record does not exist for this owner.",
        "$ref": "#/components/responses/ProblemResponse"
      },
...
      "UnprocessableEntity": {
        "description": "The request is syntactically valid but violates Jarvis validation or safety rules.",
        "$ref": "#/components/responses/ProblemResponse"
      },
...
      "ServiceUnavailable": {
        "description": "A required persistence, reliability, or safety dependency is unavailable.",
        "$ref": "#/components/responses/ProblemResponse"
      },
      "InternalServerError": {
        "description": "Jarvis failed without exposing credentials or internal exception details.",
        "$ref": "#/components/responses/ProblemResponse"
      },
      "ProblemResponse": {
        "description": "A problem response.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    },
    "schemas": {
      "RecordId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 256,
        "pattern": "^[^\\s]+$",
        "description": "Opaque identifier. Clients must not parse provider-specific structure."
      },
      "TimestampMs": {
        "type": "integer",
        "format": "int64",
        "minimum": 0,
        "description": "Unix timestamp in milliseconds."
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "instance", "requestId"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "description": "Stable problem type identifier."
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "Safe operator-facing detail with credentials and internal exception data removed."
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["path", "message"],
        "properties": {
          "path": {
            "type": "string",
            "maxLength": 500
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        }
      },
...
      "Property": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "clientId", "address", "hazards", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RecordId"
          },
          "clientId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "hazards": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "accessNotes": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "serviceNotes": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "createdAt": {
            "$ref": "#/components/schemas/TimestampMs"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/TimestampMs"
          }
        }
      },
      "PropertyResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data"],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Property"
          }
        }
      },
      "PropertyListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data", "count"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Property"
            }
          },
          "count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "CreatePropertyRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["clientId", "address"],
        "properties": {
          "clientId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "hazards": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "accessNotes": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "serviceNotes": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        }
      },
      "UpdatePropertyRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "clientId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "hazards": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "accessNotes": {
            "type": ["string", "null"],
            "maxLength": 2000
          },
          "serviceNotes": {
            "type": ["string", "null"],
            "maxLength": 2000
          }
        }
      },
...
    }
  }
}